Move tooltip event to EventListener.java
